migrate test_uptime to pytest
